Adds x-versions of mimetypes to response formatter.
After the commit updating the accepted mimetypes at image add,
the two lists of accepted mimetypes are no longer in sync. If an image
without an extension is requested, Imbo will generate an error on
retrieval unless the corresponding mimetypes are added to the
response formatter as well.

We might want to work this into one single list somewhere down the line.