Add a histogram transformation.
This transformation converts the image into a graph representation
of the color usage in the image. The base size of the histogram is
a width of 256px.

The transformation supports four parameters:

    scale: Allows you to set the width of each bar in the histogram.
           Setting a large scale will allow for anti-aliasing later.
      red: The color to use when drawing the red part of the graph.
    green: Same for green.
     blue: .. and same for blue.

Add a resize/maxSize transformation to resize the resulting image
down to the wanted size if you use the scale parameter and don't want a
size multiplied with 256.