Fix usage of hard coded temp directory in FileSystem-based tests.
There are two tests that attempt to create files in the system's temp
directory, but uses a hard coded path to /tmp to do so. This patch changes
the tests to use sys_get_temp_dir() to retrieve the actual location
instead.

Since I didn't want to override the PHPUnit Test Case constructor for
compability reasons, I decided to initialize the variable in the setUp
method instead.