Add basic support for handling PDFs in a Vector EventListener.
The listener intercepts storage and retrieval of the images, converting
the image to a raster version upon retrieval. This allows us to keep the
vector file in storage, while generating rasterized versions when a
request is made.

This is currently borked when storing files, as the Imagick-listener
apparently needs to be able to re-read the imageblob, so future patches
will solve that issue. We'll need to discuss how to handle this properly
for future implementation as well.

For now, comment out line 86,
`$this->imagick->readImageBlob($image->getBlob());` in the Imagick event
listener to store your vector file, then move it back in again.