Remove useless test
In the scenario title it said request shortUrl with no public key or access token provided, but both the public key and access token was previosly set as part of the "I generate a short URL with the following parameters" step. The adding of shortUrls is tested in another feature.