Add test for Doctrine exception on failed insert
Sorry about the mocks.

This was the only decent way of testing that the code actually throws an
exception, as I have not found a decent way of making the insert fail in
an integration test.

Only the `imageExists` method inside the actual Doctrine implementation
is mocked to avoid having to implement all the details for the connection.