#428: Validate signed URLs regardless of t[] vs t[0] (#475)
* Change mock expectation to atLeastOnce()

The previous implementation assumed that we'd only want to retrieve our URLs once, but we'd prefer to read them several times.

* Add test for t[] vs t[0] handling for accessTokens

* Add helper class for flattening arrays in query strings

* #428: Fix issue with t[] vs t[0] in URLs breaking accessToken

As detailed in Issue #428, certain sites (Facebook) regenerates URLs on their side where arguments using the array syntax of PHP gets an index appended in the URL. This breaks our signed requests and ends up with Facebook receiving a 400 HTTP error response instead of the image.

By using the Http\QueryString class from Guzzle we can work around the issue by introducing two new URLs to the list of URLs we test when we check the signature. These URLs will have t[] replaced with t[0] (and up), or t[0], t[1], etc. replaced with t[]. This way we'll accept both possible changes.

In cases where parameters are intermingled (t[]=foo&b[]=bar&t[]=baz) this will not help, but I'm fairly certain we won't see any requests like that in the wild. In addition any request made as t[1]=foo&t[0]=bar is undefined (and might give unexpected results).

* Change helper class name and add build URL method

* Add guzzlehttp/psr7 dependency and update lockfile

* Rewrite index array converter for guzzlehttp\psr7.

The old version had been removed and deprecated, but by adding a helper
function to use `parse_url` data (which is what the Url class in
guzzlehttp\psr7 does internally), we can work around limitations in the
current one (as it _always_ escapes [] in query argument names, which is
suitable for URLs to be used, but not for us - we're validating the
content for the signature).

* phpdoc-fix