Wmark: Replace setImageOpacity with evaluateImage
The previous method of using ->setImageOpacity doesn't play nice with
images with an existing alpha channel, so we use evaluateImage instead to
properly multiply the existing alpha values. That way any transparent area
in an image will remain transparent even after being composited into the
original image as a watermark.