Fix config variable mangling
If a local config file used the same variable internally as the one used
in config.default.php, the internal values was overwritten by the name in
the local configuration file. This meant we ended up with a cryptic errors
about missing routes or other default values that were not initialized as
they should have been.

By wrapping the require in a closure, we keep any local variable declared
internal to the closure and thus avoid introducing them to the scope of
the default configuration file.