Move loaders / output converters to default config
Instead of having the basic input loader and the basic output converter
registered in the input loader manager the output converter manager by
hard coding it in the application class we want it in the default
configuration instead. The values in the list will need keys for
identification for merging of custom configuration to work as expected.