Add 8BIM profile path extraction event listener
This event listener extracts paths from the 8BIM metadata and adds them
as metadata under the "paths" element in the metadata store.

It currently only extracts the paths, but if necessary in the future
can be expanded to also extract and parse additional fields. In that
cases we should consider a separate "8bim:" namespace for metadata.