Apply and track requested output compression in image model
Since the output converters now has the final say in how an image should
be rendered, and the old compress transformation was borked by the
introduction of the transformationManager, we now explicitly track the
requested compression quality on the image model.

This allows to apply the compression quality selectively on output
instead of having to try to work around when the different classes are
created and registered for transformations and events.